% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GNLHD_Construction .R
\docType{data}
\name{GGNLHD}
\alias{GGNLHD}
\title{GGNLHD class}
\format{\preformatted{Class 'R6ClassGenerator' <environment: 0x0f8cc468> 
 - attr(*, "name")= chr "GGNLHD_generator"
}}
\usage{
GGNLHD
}
\description{
GGNLHD class
}
\keyword{datasets}

