% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GNLHD_Construction .R
\docType{data}
\name{GNLHD}
\alias{GNLHD}
\title{GNLHD class}
\format{\preformatted{Class 'R6ClassGenerator' <environment: 0x0f8b1ab0> 
 - attr(*, "name")= chr "GNLHD_generator"
}}
\usage{
GNLHD
}
\description{
GNLHD class
}
\keyword{datasets}

