% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GNLHD_Construction .R
\docType{data}
\name{LHD}
\alias{LHD}
\title{LHD class}
\format{\preformatted{Class 'R6ClassGenerator' <environment: 0x0f8c2920> 
 - attr(*, "name")= chr "LHD_generator"
}}
\usage{
LHD
}
\description{
LHD class
}
\keyword{datasets}

