% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GNLHD_Construction .R
\docType{data}
\name{NLHD}
\alias{NLHD}
\title{NLHD class}
\format{\preformatted{Class 'R6ClassGenerator' <environment: 0x0f8ba124> 
 - attr(*, "name")= chr "NLHD_generator"
}}
\usage{
NLHD
}
\description{
NLHD class
}
\keyword{datasets}

