% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Sequential_Optimizing_Algorithm.R
\name{Optimal_GNLHD_SequentialAlg}
\alias{Optimal_GNLHD_SequentialAlg}
\title{Optimal_GNLHD_SequentialAlg function}
\usage{
Optimal_GNLHD_SequentialAlg(GNLHD, GNLH_Full, iteration)
}
\description{
Optimal_GNLHD_SequentialAlg function
}

