% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Plot_Grid.R
\name{Phi_p_distribution}
\alias{Phi_p_distribution}
\title{Phi_p_distribution function}
\usage{
Phi_p_distribution(structure, dimension, number, p_value, t_value)
}
\description{
Phi_p_distribution function
}

